Use nerves_ssh for providing ssh support
This refactors SSH support to its own library so that projects not using
`nerves_pack` don't have to copy/paste the SSH code. `nerves_ssh` also
has several improvements to SSH support that will be easier to maintain
in a standalone project.

This also brings in the switch from `nerves_firmware_ssh` to
`ssh_subsystem_fwup` which moves firmware updates from port 8989 to an
SSH subsystem on port 22. This is a breaking change for scripts that
load firmware to Nerves devices via SSH. `mix firmware.gen.script`
creates an updated script, but care should be taken to use the right
script based on the version of `nerves_pack` running on the device. It's
also possible to run a firmware update via `mix upload`.