Work around mdns_lite's optional dep not working
This seems to fix an issue where `:mdns_lite` is started before
`:vintage_net` in the OTP release. When this happens, `:mdns_lite`'s
Application.start fails and terminates initialization. This is due to
dependency order not being respected on optional dependencies.

See https://github.com/erlang/otp/pull/2675 for the work to fix this.