Remove automatic mdns_lite setup
This wasn't documented and since we've removed all of the other
automatic configuration setup code, it seems consistent to remove this
too.

A nice side effect of this is that this project becomes a dependency and
documentation project. IMHO, it looks much more approachable than
`nerves_init_gadget` ever did when the time came for a project to move
beyond the defaults setup.