Require constituant libraries to enforce versions
The value of this library is that it saves time adding the base set of
Nerves dependencies to projects. If there's a problem with using certain
versions of libraries then that should be sorted out in the library
because not everyone will use `nerves_pack`. Specifying them here adds
to a step to maintenance.

This also removes dependabot. While I think there's value in testing
new library combinations. Dependabot would make updates that disallowed
old versions on major updates. The old versions still worked and this
could easily have introduced a unnecessary dependency version conflict.